+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Css is broken on statistics module ([https://github.com/GotCms/GotCms/issues/166](https://github.com/GotCms/GotCms/issues/166))
 - Photoartwork theme not working properly ([https://github.com/GotCms/GotCms/issues/168](https://github.com/GotCms/GotCms/issues/168))
 - Change some confirm box into jQuery ui Dialog ([https://github.com/GotCms/GotCms/issues/167](https://github.com/GotCms/GotCms/issues/167))
+- Add parameter to force the branch to be active on Gc\Component\Navigation ([https://github.com/GotCms/GotCms/issues/169](https://github.com/GotCms/GotCms/issues/169))
 
 ##1.1.0 (11 Dec 2013)
 - Update to Zend Framework 2.2.5 ([https://github.com/GotCms/GotCms/issues/141](https://github.com/GotCms/GotCms/issues/141))